move common code to new base class
