+++ fastos/src/tests/threadtest.cpp
@@ -4,64 +4,50 @@
 #include <vespa/fastos/fastos.h>
 #include "tests.h"
 #include "jobs.h"
+#include "base_thread.hpp"
 
-
-static volatile int64_t number;
-#define INCREASE_NUMBER_AMOUNT 10000
 
 #define MUTEX_TEST_THREADS 6
 #define MAX_THREADS 7
 
 
-class ThreadTest : public BaseTest, public FastOS_Runnable
+class ThreadTest : public BaseForThreadTest
 {
-private:
-   FastOS_Mutex printMutex;
+   int Main ();
 
-public:
-   ThreadTest(void)
-     : printMutex()
+   void WaitForXThreadsToHaveWait (Job *jobs,
+                                   int jobCount,
+                                   FastOS_Cond *condition,
+                                   int numWait)
    {
-   }
-   virtual ~ThreadTest() {};
+      Progress(true, "Waiting for %d threads to be in wait state", numWait);
 
-   void PrintProgress (char *string)
-   {
-      printMutex.Lock();
-      BaseTest::PrintProgress(string);
-      printMutex.Unlock();
-   }
-
-   void Run (FastOS_ThreadInterface *thread, void *arg);
-
-   void WaitForThreadsToFinish (Job *jobs, int count)
-   {
-      int i;
-
-      Progress(true, "Waiting for threads to finish...");
+      int oldNumber=-10000;
       for(;;)
       {
-         bool threadsFinished=true;
+         int waitingThreads=0;
 
-         for(i=0; i<count; i++)
+         condition->Lock();
+
+         for(int i=0; i<jobCount; i++)
          {
-            if(jobs[i].result == -1)
-            {
-               threadsFinished = false;
-               break;
-            }
+            if(jobs[i].result == 1)
+               waitingThreads++;
          }
 
-         FastOS_Thread::Sleep(500);
+         condition->Unlock();
 
-         if(threadsFinished)
+         if(waitingThreads != oldNumber)
+            Progress(true, "%d threads are waiting", waitingThreads);
+
+         oldNumber = waitingThreads;
+
+         if(waitingThreads == numWait)
             break;
+
+         FastOS_Thread::Sleep(100);
       }
-
-      Progress(true, "Threads finished");
    }
-
-   int Main ();
 
    void TooManyThreadsTest ()
    {
@@ -337,40 +323,6 @@ public:
       PrintSeparator();
    }
 
-   void WaitForXThreadsToHaveWait (Job *jobs,
-                                   int jobCount,
-                                   FastOS_Cond *condition,
-                                   int numWait)
-   {
-      Progress(true, "Waiting for %d threads to be in wait state", numWait);
-
-      int oldNumber=-10000;
-      for(;;)
-      {
-         int waitingThreads=0;
-
-         condition->Lock();
-
-         for(int i=0; i<jobCount; i++)
-         {
-            if(jobs[i].result == 1)
-               waitingThreads++;
-         }
-
-         condition->Unlock();
-
-         if(waitingThreads != oldNumber)
-            Progress(true, "%d threads are waiting", waitingThreads);
-
-         oldNumber = waitingThreads;
-
-         if(waitingThreads == numWait)
-            break;
-
-         FastOS_Thread::Sleep(100);
-      }
-   }
-
    void SharedSignalAndBroadcastTest (Job *jobs, int numThreads,
                                       FastOS_Cond *condition,
                                       FastOS_ThreadPool *pool)
@@ -642,172 +594,6 @@ int ThreadTest::Main ()
    printf("END OF TEST (%s)\n", _argv[0]);
 
    return 0;
-}
-
-volatile int busyCnt;
-
-void ThreadTest::Run (FastOS_ThreadInterface *thread, void *arg)
-{
-   if(arg == NULL)
-      return;
-
-   Job *job = static_cast<Job *>(arg);
-   char someStack[15*1024];
-
-   memset(someStack, 0, 15*1024);
-
-   switch(job->code)
-   {
-      case SILENTNOP:
-      {
-         job->result = 1;
-         break;
-      }
-
-      case NOP:
-      {
-         Progress(true, "Doing NOP");
-         job->result = 1;
-         break;
-      }
-
-      case PRINT_MESSAGE_AND_WAIT3SEC:
-      {
-         Progress(true, "Thread printing message: [%s]", job->message);
-         job->result = strlen(job->message);
-
-         FastOS_Thread::Sleep(3000);
-         break;
-      }
-
-      case INCREASE_NUMBER:
-      {
-         int result;
-
-         if(job->mutex != NULL)
-            job->mutex->Lock();
-
-         result = static_cast<int>(number);
-
-         int sleepOn = (INCREASE_NUMBER_AMOUNT/2) * 321/10000;
-         for(int i=0; i<(INCREASE_NUMBER_AMOUNT/2); i++)
-         {
-            number = number + 2;
-
-            if(i == sleepOn)
-               FastOS_Thread::Sleep(1000);
-         }
-
-         if(job->mutex != NULL)
-            job->mutex->Unlock();
-
-         job->result = result;  // This marks the end of the thread
-
-         break;
-      }
-
-      case WAIT_FOR_BREAK_FLAG:
-      {
-         for(;;)
-         {
-            FastOS_Thread::Sleep(1000);
-
-            if(thread->GetBreakFlag())
-            {
-               Progress(true, "Thread %p got breakflag", thread);
-               break;
-            }
-         }
-         break;
-      }
-
-      case WAIT_FOR_THREAD_TO_FINISH:
-      {
-         if(job->mutex)
-            job->mutex->Lock();
-
-         if(job->otherThread != NULL)
-            job->otherThread->Join();
-
-         if(job->mutex)
-            job->mutex->Unlock();
-         break;
-      }
-
-      case WAIT_FOR_CONDITION:
-      {
-         job->condition->Lock();
-
-         job->result = 1;
-
-         job->condition->Wait();
-         job->condition->Unlock();
-
-         job->result = 0;
-
-         break;
-      }
-
-      case BOUNCE_CONDITIONS:
-      {
-        while (!thread->GetBreakFlag()) {
-          job->otherjob->condition->Lock();
-          job->otherjob->bouncewakeupcnt++;
-          job->otherjob->bouncewakeup = true;
-          job->otherjob->condition->Signal();
-          job->otherjob->condition->Unlock();
-
-          job->condition->Lock();
-          while (!job->bouncewakeup)
-            job->condition->TimedWait(1);
-          job->bouncewakeup = false;
-          job->condition->Unlock();
-        }
-        break;
-      }
-
-      case TEST_ID:
-      {
-         job->mutex->Lock();          // Initially the parent threads owns the lock
-         job->mutex->Unlock();        // It is unlocked when we should start
-
-         FastOS_ThreadId currentId = FastOS_Thread::GetCurrentThreadId();
-
-         if(currentId == job->_threadId)
-            job->result = 1;
-         else
-            job->result = -1;
-         break;
-      }
-
-      case WAIT2SEC_AND_SIGNALCOND:
-      {
-         FastOS_Thread::Sleep(2000);
-         job->condition->Signal();
-         job->result = 1;
-         break;
-      }
-
-      case HOLD_MUTEX_FOR2SEC:
-      {
-         job->mutex->Lock();
-         FastOS_Thread::Sleep(2000);
-         job->mutex->Unlock();
-         job->result = 1;
-         break;
-      }
-
-      case WAIT_2_SEC:
-      {
-         FastOS_Thread::Sleep(2000);
-         job->result = 1;
-         break;
-      }
-
-      default:
-         Progress(false, "Unknown jobcode");
-         break;
-   }
 }
 
 int main (int argc, char **argv)
